--- include/GameState.h
@@ -1,14 +1,0 @@
-#ifndef GAMESTATE_H
-#define GAMESTATE_H
-
-namespace GS {
-
-    enum GameState                 //перечисление сосостояния
-    {
-        MENU,
-        BOARD,
-        EXIT
-    };
-}
-
-#endif // GAMESTATE_H
