+++ include/palette.hpp
