+++ src/Menu.cpp
@@ -1,5 +1,5 @@
-#include "Menu.h"
-#include "palette.h"
+#include "Menu.hpp"
+#include "palette.hpp"
 #include <curses.h>
 
 Menu::Menu() {
